security-platform/pyproject.toml
[project]
name = "security-platform"
version = "0.1.0"
description = "Agent Security Platform - MCP/Skill/Agent security monitoring"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "apscheduler>=3.10.0",
    "jinja2>=3.1.4",
    "anthropic>=0.30.0",
    "google-generativeai>=0.8.0",
    "rich>=13.7.0",
    "typer>=0.12.0",
]

[project.scripts]
security-platform = "src.dashboard.app:main"
security-collector = "src.collector.main:main"
security-analyzer = "src.analyzer.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
anyio_backend = "asyncio"

[tool.uv]
dev-dependencies = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "anyio[trio]>=4.4.0",
    "httpx>=0.27.0",
]
